+++ controllers/missionsController.js
@@ -12,25 +12,38 @@ class MissionsController {
     } catch (error) {
       const message = `Error in getMissions controller: ${error.message}`;
       console.error(message);
-      throw new Error({ error: error.message });
+      res.status(500).json({ error: "INternal server error" });
     }
   }
 
   async getMissionById(req, res) {
     const { id } = req.params;
+    if (!id) {
+      return res.status(400).json({ error: "ID required" });
+    }
     try {
       const missionById = await this.missionsService.getMissionById(id);
+      if (!missionById || missionById.length === 0) {
+        return res.status(404).json({ error: "MIssion not found" });
+      }
       res.status(200).json({ missionById });
     } catch (error) {
       const message = `Error un getMissionById controller: ${error.message}`;
       console.error(message);
-      throw new Error(message);
+      res.status(500).json({ error: "Internal server error" });
     }
   }
 
   async getApplicationByMissionId(req, res) {
     const { id } = req.params;
+    if (!id) {
+      return res.status(400).json({ error: "Id required" });
+    }
     try {
+      const mission = await this.missionsService.getMissionById(id);
+      if (!mission || mission.length === 0) {
+        return res.status(404).json({ error: "mission not found" });
+      }
       const applications = await this.missionsService.getApplicationByMissionId(
         id
       );
@@ -39,12 +52,15 @@ class MissionsController {
     } catch (error) {
       const message = `Error in getApplicationByMissionId controller: ${error.message}`;
       console.error(message);
-      res.status(404).json({ error: "Application not found" });
+      res.status(500).json({ error: "internal server error" });
     }
   }
 
   async createMission(req, res) {
     const { title, description, mission_date, association_id } = req.body;
+    if (!title || !description || !mission_date || !association_id) {
+      return res.status(400).json({ error: "Missing required fileds" });
+    }
 
     try {
       await this.missionsService.createMission(
@@ -57,43 +73,53 @@ class MissionsController {
     } catch (error) {
       const message = `Error in createMission controller: ${error.message}`;
       console.error(message);
-      res.status(400).json({ error: error.message });
+
+      res.status(500).json({ error: "INternal server error" });
     }
   }
 
   async updateMissionById(req, res) {
     const { id } = req.params;
     const { title, description, mission_date } = req.body;
+
+    if (!id || !title || !description || !mission_date) {
+      return res.status(400).json({ error: " missing one required field" });
+    }
     try {
-      const missionUpdate = await this.missionsService.updateMissionById(
+      const missionUpdated = await this.missionsService.updateMissionById(
         id,
         title,
         description,
         mission_date
       );
-      if (missionUpdate) {
-        return res
-          .status(200)
-          .json({ message: "Mission updated with success" });
-      } else {
-        return res.status(404).json({ error: "Mission not found" });
+      if (!missionUpdated || missionUpdated.length === 0) {
+        return res.status(404).json({ error: "mission not found" });
       }
+
+      res.status(200).send("MIsssion updated");
     } catch (error) {
       const message = `Error in updateMissionById controller: ${error.message}`;
       console.error(message);
-      res.status(500).json({ error: error.message });
+      res.status(500).json({ error: "Internal server error" });
     }
   }
 
   async deleteMissionById(req, res) {
     const { id } = req.params;
+    if (!id) {
+      return res.status(400).json({ error: "id required" });
+    }
+
     try {
-      await this.missionsService.deleteMissionById(id);
+      const missionDeleted = await this.missionsService.deleteMissionById(id);
+      if (!missionDeleted || missionDeleted.affectedRows === 0) {
+        return res.status(404).json({ message: "Mission not found" });
+      }
       res.status(200).json({ message: "Mission deleted with success" });
     } catch (error) {
       const message = `Error in updateMissionById controller: ${error.message}`;
       console.error(message);
-      res.status(500).json({ error: error.message });
+      res.status(500).json({ error: "Internal server error" });
     }
   }
 }
